change to floating point in intoft.c

+++ intoft.c
@@ -10,7 +10,7 @@ int main(int argc, char** argv) {
         fprintf(stderr, "Usage: %s [inches to convert to feet]\n", argv[0]);
         return 1;
     }
-    int inches = atoi(argv[1]);
-    printf("%d'%d\"", inches / 12, inches % 12);
+    double inches = atof(argv[1]);
+    printf("%d'%d\"", (int)(inches / 12), (int)(inches % 12));
     return 0;
 }
